% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiZer.R
\name{as.data.frame.SiZer}
\alias{as.data.frame.SiZer}
\title{Coerce SiZer object to a Data Frame}
\usage{
\method{as.data.frame}{SiZer}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An object produced by `SiZer()`.}

\item{row.names}{Required for generic compatibility. Not used.}

\item{optional}{Required for generic compatibility. Not used.}

\item{...}{Required for generic compatibility. Not used.}
}
\description{
Coerce SiZer object to a Data Frame
}
