% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bent.cable.R
\name{logLik.bent_cable}
\alias{logLik.bent_cable}
\title{Return the log-Likelihood value for a fitted bent-cable model.}
\usage{
\method{logLik}{bent_cable}(object, ...)
}
\arguments{
\item{object}{A bent-cable model}

\item{\dots}{Unused at this time.}
}
\description{
Return the log-Likelihood value for a fitted bent-cable model.
}
