% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.linear.R
\name{logLik.PiecewiseLinear}
\alias{logLik.PiecewiseLinear}
\title{Calculates the log-Likelihood value}
\usage{
\method{logLik}{PiecewiseLinear}(object, ...)
}
\arguments{
\item{object}{A \code{PiecewiseLinear} object}

\item{\dots}{Unused at this time.}
}
\description{
Calculates the log-Likelihood value
}
