% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone.R
\name{phone}
\alias{phone}
\title{Export a phone-sized jpeg}
\usage{
phone(filename, plot, filetype = c("jpeg", "png"), dpi = 300)
}
\arguments{
\item{filename}{the filename including the extension (currently only supports jpeg)}

\item{plot}{the plot object you want to export in the phone size}
}
\description{
Saves a plot object as a jpeg file that meets the dimension requirements of the "mobile devices and skins" product on society6.com (1300 pixels in width, 2000 pixels in height)
}
\details{
Note: current options are jpg or png, and this function has NOT been tested on Windows.
}
\examples{
# Base R example
p <- plot(iris$Sepal.Length, y = iris$Petal.Length, type = "p", col = iris$Species)
phone(filename = "iris_baseR.jpeg", plot = p, filetype = "jpeg")

## ggplot2 example
p <- ggplot(iris, aes(Sepal.Length, Petal.Length)) + geom_point(aes(color = Species))
phone(filename = "iris_ggplot.png", plot = p, filetype ="png")
}

