% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locally.weighted.polynomial.R
\name{plot.LocallyWeightedPolynomial}
\alias{plot.LocallyWeightedPolynomial}
\title{Creates a plot of an object created by \code{locally.weighted.polynomial}.}
\usage{
\method{plot}{LocallyWeightedPolynomial}(x, derv = 0, CI.method = 2,
  alpha = 0.05, use.ess = TRUE, draw.points = TRUE, ...)
}
\arguments{
\item{x}{LocallyWeightedPolynomial object}

\item{derv}{Derivative to be plotted.  Default is 0 - which plots the smoothed function.}

\item{CI.method}{What method should be used to calculate the confidence interval about the estimated line.
The methods are from Hannig and Marron (2006), where 1 is the point-wise estimate, and 2 is 
the row-wise estimate.}

\item{alpha}{The alpha level such that the CI has a 1-alpha/2 level of significance.}

\item{use.ess}{ESS stands for the estimated sample size.  If at any point along the x-axis, the ESS is 
too small, then we will not plot unless use.ess=FALSE.}

\item{draw.points}{Should the data points be included in the graph? Defaults to TRUE.}

\item{\dots}{Additional arguments to be passed to the graphing functions.}
}
\description{
Creates a plot of an object created by \code{locally.weighted.polynomial}.
}
