% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.linear.R
\name{plot.PiecewiseLinear}
\alias{plot.PiecewiseLinear}
\title{Plots a piecewise linear model}
\usage{
\method{plot}{PiecewiseLinear}(x, xlab = "X", ylab = "Y", ...)
}
\arguments{
\item{x}{A \code{PiecewiseLinear} object}

\item{xlab}{The label for the x-axis}

\item{ylab}{The label for the y-axis}

\item{\dots}{Any further options to be passed to the \code{plot} function}
}
\description{
Plots a piecewise linear model
}
