% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bent.cable.R
\name{predict.bent_cable}
\alias{predict.bent_cable}
\title{Return model predictions for fitted bent-cable model}
\usage{
\method{predict}{bent_cable}(object, x, ...)
}
\arguments{
\item{object}{A bent-cable model}

\item{x}{The set x-values for which predictions are desired}

\item{\dots}{A placeholder that is currently ignored.}
}
\description{
Return model predictions for fitted bent-cable model
}
