% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.linear.R
\name{predict.PiecewiseLinear}
\alias{predict.PiecewiseLinear}
\title{Calculates predicted values from a piecewise linear object}
\usage{
\method{predict}{PiecewiseLinear}(object, x, ...)
}
\arguments{
\item{object}{A \code{PiecewiseLinear} object}

\item{x}{A vector of x-values in which to calculate the y.}

\item{\dots}{Unused at this time.}
}
\description{
Calculates predicted values from a piecewise linear object
}
