% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.linear.R
\name{print.PiecewiseLinear}
\alias{print.PiecewiseLinear}
\title{Prints out the model form for a Piecewise linear model}
\usage{
\method{print}{PiecewiseLinear}(x, ...)
}
\arguments{
\item{x}{A \code{PiecewiseLinear} object}

\item{\dots}{Unused at this time.}
}
\description{
Prints out the model form for a Piecewise linear model
}
