% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizer.R
\docType{package}
\name{sizer}
\alias{sizer}
\alias{sizer-package}
\title{sizer: a package for exporting plot objects to pre-defined sizes of jpeg files}
\description{
sizer: a package for exporting plot objects to pre-defined sizes of jpeg files
}

