% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Export the a 6000 x 6000 px jpeg or png image}
\usage{
square(filename, plot, filetype = c("jpeg", "png"), dpi = 300)
}
\arguments{
\item{filename}{the filename including the extension}

\item{plot}{the plot object you want to export}
}
\description{
Saves a plot object as an image file that meets the dimension requirements of most products on Society6 (7500 x 9000 px)
}
\details{
Note: current options are jpg or png, and this function has NOT been tested on Windows.
}
\examples{
# Base R example
p <- plot(iris$Sepal.Length, y = iris$Petal.Length, type = "p", col = iris$Species)
generic(filename = "iris_baseR.jpeg", plot = p, filetype = "jpeg")

## ggplot2 example
p <- ggplot(iris, aes(Sepal.Length, Petal.Length)) + geom_point(aes(color = Species))
generic(filename = "iris_ggplot.png", plot = p, filetype ="png")
}

