% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomm.R
\name{tomm}
\alias{tomm}
\title{Convert pixel width to millimeters with a given dpi value}
\usage{
tomm(width, height, dpi = 300)
}
\arguments{
\item{width}{width in pixels}

\item{height}{height in pixels}

\item{dpi=300}{dpi value}
}
\value{
the new width and height values in millimeters
}
\description{
Convert pixel width to millimeters with a given dpi value
}
\examples{
tomm(1300, 2000)
}

