% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tshirt.R
\name{tshirt}
\alias{tshirt}
\title{Export a t-shirt-sized image}
\usage{
tshirt(filename = "plot", plot, filetype = c("jpeg", "png"), dpi = 300)
}
\arguments{
\item{filename}{the filename including the extension}

\item{plot}{the plot object you want to export in the t-shirt size}

\item{filetype}{choose either "jpeg" or "png"}

\item{dpi}{default is 300}
}
\description{
Saves a plot object as an image file that meets the dimension requirements of the "T-shirts/V-necks/Tank Tops/Long Sleeve/Biker Tanks/Hoodies" product on society6.com (1300 pixels in width, 2000 pixels in height)
}
\details{
Note: current options are jpg or png, and this function has NOT been tested on Windows.  If you select "png" as your filetype, note that the transparency will be set to on.
}

